refactor: ProfileScreen UI 컴포넌트 분리 및 구조 개선
`ProfileScreen`의 가독성과 유지보수성을 높이기 위해 내부 UI 로직을 별도 컴포저블로 분리하고 코드 구조를 정리했습니다.
